Implement UnderlineGrammar and refactor package
