+++ src/main/java/com/github/freenamu/parser/grammar/singleline/BoldGrammar.java
@@ -1,6 +1,7 @@
-package com.github.freenamu.parser.grammar;
+package com.github.freenamu.parser.grammar.singleline;
 
-import com.github.freenamu.parser.node.Italic;
+import com.github.freenamu.parser.grammar.LeafGrammar;
+import com.github.freenamu.parser.node.Bold;
 import com.github.freenamu.parser.node.Node;
 import com.github.freenamu.parser.node.Text;
 
@@ -9,9 +10,9 @@ import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-public class ItalicGrammar extends LeafGrammar {
-    public ItalicGrammar() {
-        super(Pattern.compile("''[^\n]+?''"));
+public class BoldGrammar extends LeafGrammar {
+    public BoldGrammar() {
+        super(Pattern.compile("'''[^\n]+?'''"));
     }
 
     @Override
@@ -21,9 +22,9 @@ public class ItalicGrammar extends LeafGrammar {
 
         if (matcher.find()) {
             result.add(new Text(rawText.substring(0, matcher.start())));
-            Italic italic = new Italic();
-            italic.addAll(new SingleLineGrammar().parse(rawText.substring(matcher.start() + 2, matcher.end() - 2)));
-            result.add(italic);
+            Bold bold = new Bold();
+            bold.addAll(new SingleLineGrammar().parse(rawText.substring(matcher.start() + 3, matcher.end() - 3)));
+            result.add(bold);
             result.addAll(new SingleLineGrammar().parse(rawText.substring(matcher.end())));
         }
