+++ src/main/java/com/github/freenamu/parser/grammar/singleline/QuoteGrammar.java
@@ -1,5 +1,6 @@
-package com.github.freenamu.parser.grammar;
+package com.github.freenamu.parser.grammar.singleline;
 
+import com.github.freenamu.parser.grammar.CompositeGrammar;
 import com.github.freenamu.parser.node.Node;
 import com.github.freenamu.parser.node.Text;
