+++ src/main/java/com/github/freenamu/parser/grammar/singleline/UnderlineGrammar.java
@@ -1,17 +1,18 @@
-package com.github.freenamu.parser.grammar;
+package com.github.freenamu.parser.grammar.singleline;
 
-import com.github.freenamu.parser.node.Bold;
+import com.github.freenamu.parser.grammar.LeafGrammar;
 import com.github.freenamu.parser.node.Node;
 import com.github.freenamu.parser.node.Text;
+import com.github.freenamu.parser.node.Underline;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-public class BoldGrammar extends LeafGrammar {
-    public BoldGrammar() {
-        super(Pattern.compile("'''[^\n]+?'''"));
+public class UnderlineGrammar extends LeafGrammar {
+    public UnderlineGrammar() {
+        super(Pattern.compile("__[^\n]+?__"));
     }
 
     @Override
@@ -21,9 +22,9 @@ public class BoldGrammar extends LeafGrammar {
 
         if (matcher.find()) {
             result.add(new Text(rawText.substring(0, matcher.start())));
-            Bold bold = new Bold();
-            bold.addAll(new SingleLineGrammar().parse(rawText.substring(matcher.start() + 3, matcher.end() - 3)));
-            result.add(bold);
+            Underline underline = new Underline();
+            underline.addAll(new SingleLineGrammar().parse(rawText.substring(matcher.start() + 2, matcher.end() - 2)));
+            result.add(underline);
             result.addAll(new SingleLineGrammar().parse(rawText.substring(matcher.end())));
         }
