+++ src/main/java/com/github/freenamu/parser/node/Node.java
@@ -43,6 +43,7 @@ public abstract class Node {
     public enum Type {
         Bold,
         Italic,
+        Underline,
         Text
     }
 }
