+++ src/main/java/com/github/freenamu/parser/node/Underline.java
@@ -1,0 +1,7 @@
+package com.github.freenamu.parser.node;
+
+public class Underline extends Node {
+    public Underline() {
+        super(Type.Underline);
+    }
+}
